+++ src/resource.ts
@@ -15,22 +15,24 @@
  */
 
 import { diag } from '@opentelemetry/api';
-import { Resource } from '@opentelemetry/resources';
+import {
+  Resource,
+  envDetectorSync,
+  hostDetectorSync,
+  osDetectorSync,
+  processDetectorSync,
+} from '@opentelemetry/resources';
 
 import { distroDetector } from './detectors/DistroDetector';
-import { dockerCGroupV1Detector } from './detectors/DockerCGroupV1Detector';
-import { envDetector } from './detectors/EnvDetector';
-import { hostDetector } from './detectors/HostDetector';
-import { osDetector } from './detectors/OSDetector';
-import { processDetector } from './detectors/ProcessDetector';
+import { containerDetector } from './detectors/ContainerDetector';
 
 const detectors = [
   distroDetector,
-  dockerCGroupV1Detector,
-  envDetector,
-  hostDetector,
-  osDetector,
-  processDetector,
+  containerDetector,
+  envDetectorSync,
+  hostDetectorSync,
+  osDetectorSync,
+  processDetectorSync,
 ];
 
 export const detect = (): Resource => {